+++ index.html
@@ -470,6 +470,8 @@
         <button type="button" class="error__button">Попробовать снова</button>
       </div>
     </template>
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type='module'></script>
+    <script src="./js/adgenerator.js" type='module'></script>
+    <script src='./js/randvalues.js' type='module'></script>
   </body>
 </html>
