+++ package.json
@@ -3,6 +3,7 @@
   "version": "25.0.0",
   "private": true,
   "description": "Личный проект «Кексобукинг» от HTML Academy",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "git@github.com:htmlacademy-javascript/1564519-keksobooking-25.git"
